+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Change Log
 ----------
 
+DEV:
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+To be announced
+
+* added two fields to asset forms ('service catalog', 'environment')
+
+
 2.0.2-rc3
 ~~~~~~~~~
 Released on July 9, 2014